Reactoring the core headers folder to match the updated format with separation for create, base, types, and utils. Match file import ordering in index, and refactoring core table to separate types
